+++ python_module/component/pipeline.py
@@ -22,9 +22,10 @@ from gi.repository import Gst,GLib
 from python_module.common.FPS import PERF_DATA
 from python_module.component.source_factory import create_source_bin , parse_media_source
 from python_module.component.rtsp_server import create_rtsp_server
-from python_module.component.probes import sink_pad_buffer_probe
+from python_module.component.probes import sink_pad_buffer_probe, osd_sink_pad_buffer_probe
 from python_module.component.pre_process import pre_process
 from python_module.common.bus_call import bus_call
+from python_module.common.utils import create_dynamic_labels
 
 
 
@@ -109,12 +110,15 @@ def create_pipeline(args, model_type):
     elements["streammux"].set_property('batch-size', number_sources)
 
     if model_type == 'det':
-        elements["pgie"].set_property('config-file-path', "/apps/deepstream-yolo-e2e/config/pgie/config_pgie_yolo_det.txt")
+        pgie_conf_file="/apps/deepstream-yolo-e2e/config/pgie/config_pgie_yolo_det.txt"
     elif model_type == 'seg':
-        elements["pgie"].set_property('config-file-path', "/apps/deepstream-yolo-e2e/config/pgie/config_pgie_yolo_seg.txt")
+        pgie_conf_file="/apps/deepstream-yolo-e2e/config/pgie/config_pgie_yolo_seg.txt"
     else:
         sys.stderr.write(f"Model Type not supported {model_type}\n")
         exit
+    
+    elements["pgie"].set_property('config-file-path', pgie_conf_file )
+
 
     elements["tracker"].set_property('tracker-width', 640)
     elements["tracker"].set_property('tracker-height', 384)
@@ -220,8 +224,10 @@ def create_pipeline(args, model_type):
         
         else:
             elements["nvosd"].link(elements["sink"])
+    
+    dynamic_labels = create_dynamic_labels(pgie_conf_file)
 
-    return pipeline , element_probe, perf_data, output_file_path
+    return pipeline , element_probe, perf_data, output_file_path, dynamic_labels
 
 
 # Function to run the pipeline
@@ -230,7 +236,7 @@ def run_pipeline(args):
     if args.output == "rtsp":
         create_rtsp_server()
 
-    pipeline, element_probe, perf_data, output_file_path  = create_pipeline(args, model_type)
+    pipeline, element_probe, perf_data, output_file_path, dynamic_labels  = create_pipeline(args, model_type)
     if not pipeline:
         sys.stderr.write("Failed to create pipeline\n")
         return
@@ -245,6 +251,13 @@ def run_pipeline(args):
         sys.stderr.write("Unable to get sink pad of nvosd\n")
         return
     elementsinkpad.add_probe(Gst.PadProbeType.BUFFER, sink_pad_buffer_probe, 0, perf_data)
+    
+    osd = element_probe.get_static_pad("sink")
+    if not osd:
+        sys.stderr.write("Unable to get sink pad of nvosd\n")
+        return
+    osd.add_probe(Gst.PadProbeType.BUFFER, osd_sink_pad_buffer_probe, 0, dynamic_labels)
+
     GLib.timeout_add(5000, perf_data.perf_print_callback)
 
     print("Starting pipeline \n")
